NORM DONE FOR PARSING YAAAAY
