+++ Makefile
@@ -11,8 +11,8 @@ MLX_LIB = $(MLX)/libmlx.a
 
 MLX_LFLAGS = -L$(MLX) -lmlx -lXext -lX11 -lm
 
-SRC_DIR = src
-SRC =   $(addprefix $(SRC_DIR)/, cub3D.c parsing.c utils.c)
+SRC_DIR = src/parsing
+SRC =   $(addprefix $(SRC_DIR)/, cub3D.c parsing.c utils.c init_struct.c free.c is_valid.c check_map.c get_texture.c line.c get_player.c map_utils.c)
 
 OBJ_DIR = obj
 OBJ = $(SRC:$(SRC_DIR)/%.c=$(OBJ_DIR)/%.o)
@@ -45,6 +45,10 @@ fclean: clean
 	$(RM) $(NAME)
 	make -C $(LIBFT) fclean
 
+mlx:
+	git clone https://github.com/42Paris/minilibx-linux.git src/mlx_linux
+	make -C $(MLX)
+
 re: fclean all
 
-.PHONY: all clean fclean re
+.PHONY: all clean fclean re mlx
