+++ src/parsing/utils.c
@@ -6,7 +6,7 @@
 /*   By: nrontey <nrontey@student.42angouleme.fr    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/03 20:24:03 by nrontey           #+#    #+#             */
-/*   Updated: 2024/09/03 22:11:32 by nrontey          ###   ########.fr       */
+/*   Updated: 2024/10/10 06:22:13 by nrontey          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,4 +20,24 @@ int	ft_strcmp(const char *s1, const char *s2)
 		s2++;
 	}
 	return (*(unsigned char *)s1 - *(unsigned char *)s2);
+}
+
+int	ft_isnumeric(char *str)
+{
+	int	i;
+
+	i = 0;
+	while (str[i])
+	{
+		if (!ft_isdigit(str[i]))
+			return (0);
+		i++;
+	}
+	return (1);
+}
+
+void	ft_copy(char *src, char *dst)
+{
+	while (src && *src)
+		*dst++ = *src++;
 }
